CF-289 move Forecast to projection package
