+++ ta4j-core/src/test/java/org/ta4j/core/indicators/forecast/projection/ForecastProjectionIndicatorTest.java
@@ -16,7 +16,6 @@ import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
-import org.ta4j.core.walkforward.PredictionSnapshot;
 
 public class ForecastProjectionIndicatorTest extends AbstractIndicatorTest<ForecastProjectionIndicator, Num> {
 
@@ -27,8 +26,7 @@ public class ForecastProjectionIndicatorTest extends AbstractIndicatorTest<Forec
     @Test
     public void projectionMethodsReturnPointIndicators() {
         BarSeries series = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(1, 2, 3).build();
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(numOf(1), numOf(3)), List.of(0.05, 0.5, 0.95));
+        Forecast<Num> forecast = Forecast.ofSamples(2, 1, List.of(numOf(1), numOf(3)), List.of(0.05, 0.5, 0.95));
         ForecastProjectionIndicator indicator = new FixedForecastIndicator(series, 1, Map.of(2, forecast));
 
         assertNumEquals(2, indicator.mean().getValue(2));
@@ -40,8 +38,7 @@ public class ForecastProjectionIndicatorTest extends AbstractIndicatorTest<Forec
     @Test
     public void missingQuantileProjectionReturnsNaN() {
         BarSeries series = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(1, 2, 3).build();
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(numOf(1), numOf(3)), List.of(0.5));
+        Forecast<Num> forecast = Forecast.ofSamples(2, 1, List.of(numOf(1), numOf(3)), List.of(0.5));
         ForecastProjectionIndicator indicator = new FixedForecastIndicator(series, 0, Map.of(2, forecast));
 
         assertTrue(indicator.quantile(0.95).getValue(2).isNaN());
@@ -51,18 +48,17 @@ public class ForecastProjectionIndicatorTest extends AbstractIndicatorTest<Forec
 
         private final BarSeries series;
         private final int unstableBars;
-        private final Map<Integer, PredictionSnapshot.Forecast<Num>> values;
+        private final Map<Integer, Forecast<Num>> values;
 
-        private FixedForecastIndicator(BarSeries series, int unstableBars,
-                Map<Integer, PredictionSnapshot.Forecast<Num>> values) {
+        private FixedForecastIndicator(BarSeries series, int unstableBars, Map<Integer, Forecast<Num>> values) {
             this.series = series;
             this.unstableBars = unstableBars;
             this.values = values;
         }
 
         @Override
-        public PredictionSnapshot.Forecast<Num> getValue(int index) {
-            return values.getOrDefault(index, PredictionSnapshot.Forecast.unstable(index, 1, NaN.NaN));
+        public Forecast<Num> getValue(int index) {
+            return values.getOrDefault(index, Forecast.unstable(index, 1, NaN.NaN));
         }
 
         @Override
