+++ ta4j-core/src/test/java/org/ta4j/core/indicators/forecast/projection/ForecastTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package org.ta4j.core.walkforward;
+package org.ta4j.core.indicators.forecast.projection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -14,13 +14,13 @@ import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-class PredictionSnapshotTest {
+class ForecastTest {
 
     private static final NumFactory NUM_FACTORY = DoubleNumFactory.getInstance();
 
     @Test
     void forecastSummarizesSamplesWithInterpolatedQuantiles() {
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(4, 2,
+        Forecast<Num> forecast = Forecast.ofSamples(4, 2,
                 List.of(NUM_FACTORY.numOf(1), NUM_FACTORY.numOf(2), NUM_FACTORY.numOf(3), NUM_FACTORY.numOf(4)),
                 List.of(0.25, 0.5, 0.75));
 
@@ -36,8 +36,8 @@ class PredictionSnapshotTest {
 
     @Test
     void forecastSkipsInvalidSamplesAndSortsQuantiles() {
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(NaN.NaN, NUM_FACTORY.numOf(3), NUM_FACTORY.numOf(1)), List.of(0.95, 0.05));
+        Forecast<Num> forecast = Forecast.ofSamples(2, 1, List.of(NaN.NaN, NUM_FACTORY.numOf(3), NUM_FACTORY.numOf(1)),
+                List.of(0.95, 0.05));
 
         assertThat(forecast.sampleCount()).isEqualTo(2);
         assertThat(forecast.quantiles().keySet()).containsExactly(0.05, 0.95);
@@ -46,10 +46,10 @@ class PredictionSnapshotTest {
 
     @Test
     void forecastMapsSummaryValues() {
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(NUM_FACTORY.numOf(1), NUM_FACTORY.numOf(3)), List.of(0.5));
+        Forecast<Num> forecast = Forecast.ofSamples(2, 1, List.of(NUM_FACTORY.numOf(1), NUM_FACTORY.numOf(3)),
+                List.of(0.5));
 
-        PredictionSnapshot.Forecast<Num> doubled = forecast.map(value -> value.multipliedBy(NUM_FACTORY.numOf(2)));
+        Forecast<Num> doubled = forecast.map(value -> value.multipliedBy(NUM_FACTORY.numOf(2)));
 
         assertThat(doubled.mean()).isEqualByComparingTo(NUM_FACTORY.numOf(4));
         assertThat(doubled.median()).isEqualByComparingTo(NUM_FACTORY.numOf(4));
@@ -60,8 +60,8 @@ class PredictionSnapshotTest {
 
     @Test
     void forecastReturnsNullForMissingQuantiles() {
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(NUM_FACTORY.numOf(1), NUM_FACTORY.numOf(3)), List.of(0.5));
+        Forecast<Num> forecast = Forecast.ofSamples(2, 1, List.of(NUM_FACTORY.numOf(1), NUM_FACTORY.numOf(3)),
+                List.of(0.5));
 
         assertThat(forecast.hasQuantile(0.95)).isFalse();
         assertThat(forecast.quantile(0.95)).isNull();
@@ -69,13 +69,12 @@ class PredictionSnapshotTest {
 
     @Test
     void forecastValidatesInputs() {
-        assertThrows(IllegalArgumentException.class, () -> PredictionSnapshot.Forecast.unstable(0, 0));
+        assertThrows(IllegalArgumentException.class, () -> Forecast.unstable(0, 0));
         assertThrows(IllegalArgumentException.class,
-                () -> PredictionSnapshot.Forecast.ofSamples(0, 1, List.of(NUM_FACTORY.one()), List.of(-0.1)));
+                () -> Forecast.ofSamples(0, 1, List.of(NUM_FACTORY.one()), List.of(-0.1)));
         assertThrows(IllegalArgumentException.class,
-                () -> PredictionSnapshot.Forecast.ofSamples(0, 1, List.of(NUM_FACTORY.one()), List.of()));
-        PredictionSnapshot.Forecast<Num> forecast = PredictionSnapshot.Forecast.ofSamples(0, 1,
-                List.of(NUM_FACTORY.one()), List.of(0.5));
+                () -> Forecast.ofSamples(0, 1, List.of(NUM_FACTORY.one()), List.of()));
+        Forecast<Num> forecast = Forecast.ofSamples(0, 1, List.of(NUM_FACTORY.one()), List.of(0.5));
         assertThrows(IllegalArgumentException.class, () -> forecast.hasQuantile(-0.1));
         assertThrows(IllegalArgumentException.class, () -> forecast.quantile(1.1));
     }
