+++ ta4j-core/src/test/java/org/ta4j/core/indicators/forecast/projection/ForwardForecastIndicatorTest.java
@@ -17,7 +17,6 @@ import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
-import org.ta4j.core.walkforward.PredictionSnapshot;
 
 public class ForwardForecastIndicatorTest extends AbstractIndicatorTest<ForwardForecastIndicator, Num> {
 
@@ -28,10 +27,9 @@ public class ForwardForecastIndicatorTest extends AbstractIndicatorTest<ForwardF
     @Test
     public void reducesDistribution() {
         BarSeries series = new MockBarSeriesBuilder().withNumFactory(numFactory).withData(1, 2, 3).build();
-        PredictionSnapshot.Forecast<Num> forecastSummary = PredictionSnapshot.Forecast.ofSamples(2, 1,
-                List.of(numOf(1), numOf(3)));
+        Forecast<Num> forecastSummary = Forecast.ofSamples(2, 1, List.of(numOf(1), numOf(3)));
         ForecastProjectionIndicator forecast = new FixedForecastIndicator(series, 1, Map.of(2, forecastSummary));
-        ForwardForecastIndicator median = new ForwardForecastIndicator(forecast, PredictionSnapshot.Forecast::median);
+        ForwardForecastIndicator median = new ForwardForecastIndicator(forecast, Forecast::median);
         ForwardForecastIndicator p95 = new ForwardForecastIndicator(forecast, value -> value.quantile(0.95));
 
         assertEquals(1, median.getCountOfUnstableBars());
@@ -44,18 +42,17 @@ public class ForwardForecastIndicatorTest extends AbstractIndicatorTest<ForwardF
 
         private final BarSeries series;
         private final int unstableBars;
-        private final Map<Integer, PredictionSnapshot.Forecast<Num>> values;
+        private final Map<Integer, Forecast<Num>> values;
 
-        private FixedForecastIndicator(BarSeries series, int unstableBars,
-                Map<Integer, PredictionSnapshot.Forecast<Num>> values) {
+        private FixedForecastIndicator(BarSeries series, int unstableBars, Map<Integer, Forecast<Num>> values) {
             this.series = series;
             this.unstableBars = unstableBars;
             this.values = values;
         }
 
         @Override
-        public PredictionSnapshot.Forecast<Num> getValue(int index) {
-            return values.getOrDefault(index, PredictionSnapshot.Forecast.unstable(index, 1, NaN.NaN));
+        public Forecast<Num> getValue(int index) {
+            return values.getOrDefault(index, Forecast.unstable(index, 1, NaN.NaN));
         }
 
         @Override
